+++ frontend/micro-ui/web/micro-ui-internals/packages/modules/dristi/src/components/AddressBailBond.js
@@ -2,8 +2,20 @@ import { CardLabel, CardLabelError, LabelFieldPair, TextInput } from "@egovernme
 import React, { useMemo } from "react";
 import { formatAddress } from "../Utils";
 
-
-const AddressBailBond = ({ t, config, onSelect, formData = {}, errors, formState, control, watch, register, setError, clearErrors }) => {
+const AddressBailBond = ({
+  t,
+  config,
+  onSelect,
+  formData = {},
+  errors,
+  formState,
+  control,
+  watch,
+  register,
+  setError,
+  clearErrors,
+  formDisbalityCount,
+}) => {
   const configKey = `${config.key}`;
 
   const { inputs } = useMemo(() => {
@@ -62,7 +74,7 @@ const AddressBailBond = ({ t, config, onSelect, formData = {}, errors, formState
                         }
                         setValue2(value, input.name, input?.autoFill);
                       }}
-                      disable={input?.isDisabled}
+                      disable={input?.isDisabled || formDisbalityCount}
                       isRequired={input?.validation?.isRequired}
                       pattern={input?.validation?.pattern}
                       errMsg={input?.validation?.errMsg}
