+++ frontend/micro-ui/web/micro-ui-internals/packages/modules/dristi/src/components/SelectCustomDragDrop.js
@@ -32,7 +32,7 @@ const DragDropJSX = ({ t, currentValue, error }) => {
   );
 };
 
-function SelectCustomDragDrop({ t, config, formData = {}, onSelect, errors, setError, clearErrors }) {
+function SelectCustomDragDrop({ t, config, formData = {}, onSelect, errors, setError, clearErrors, formDisbalityCount = false }) {
   const toast = useToast();
   const inputs = useMemo(
     () =>
@@ -151,13 +151,13 @@ function SelectCustomDragDrop({ t, config, formData = {}, onSelect, errors, setE
                 t={t}
                 uploadErrorInfo={fileErrors[index]}
                 input={input}
-                disableUploadDelete={config?.disable}
+                disableUploadDelete={config?.disable || formDisbalityCount}
               />
             ))}
 
             <div className={`file-uploader-div-main ${showFileUploader ? "show-file-uploader" : ""}`}>
               <FileUploader
-                disabled={config?.disable}
+                disabled={config?.disable || formDisbalityCount}
                 handleChange={(data) => {
                   handleChange(data, input);
                 }}
