+++ frontend/micro-ui/web/micro-ui-internals/packages/modules/dristi/src/hooks/index.js
@@ -104,6 +104,7 @@ export const Urls = {
     processProfileRequest: "/case/v2/profilerequest/process",
     etreasuryCreateDemand: "/etreasury/payment/v1/_createDemand",
     taskSearch: "/task/v1/search",
+    searchBailBonds: "/bail-bond/v1/_search",
   },
   case: {
     addWitness: "/case/v1/add/witness",
