+++ frontend/micro-ui/web/micro-ui-internals/packages/modules/dristi/src/pages/employee/AdmittedCases/AdmittedCaseV2.js
@@ -55,6 +55,7 @@ import MarkAsEvidence from "./MarkAsEvidence";
 import AddWitnessModal from "@egovernments/digit-ui-module-hearings/src/pages/employee/AddWitnessModal";
 import WitnessDrawerV2 from "./WitnessDrawerV2";
 import WitnessDepositionDocModal from "./WitnessDepositionDocModal";
+import { convertTaskResponseToPayload } from "@egovernments/digit-ui-module-orders/src/utils";
 const stateSla = {
   SCHEDULE_HEARING: 3 * 24 * 3600 * 1000,
   NOTICE: 3 * 24 * 3600 * 1000,
@@ -2367,135 +2368,6 @@ const AdmittedCaseV2 = () => {
     }
   };
 
-  const onSubmit = async () => {
-    switch (primaryAction.action) {
-      case "REGISTER":
-        break;
-      case "ADMIT":
-        if ((isDelayApplicationPending || isDelayCondonationApplicable) && !isDelayApplicationCompleted) {
-          setIsOpenDCA(true);
-        } else {
-          setSubmitModalInfo({ ...admitCaseSubmitConfig, caseInfo: caseInfo });
-          setModalInfo({ type: "admitCase", page: 0 });
-          setShowModal(true);
-        }
-        break;
-      case "ISSUE_ORDER":
-        const { hearingDate, hearingNumber } = await getHearingData();
-        if (hearingNumber) {
-          const {
-            list: [orderData],
-          } = await Digit.ordersService.searchOrder({
-            tenantId,
-            criteria: { filingNumber, applicationNumber: "", cnrNumber, status: OrderWorkflowState.DRAFT_IN_PROGRESS, hearingNumber: hearingNumber },
-            pagination: { limit: 1, offset: 0 },
-          });
-          if (
-            (orderData?.orderCategory === "COMPOSITE" && orderData?.compositeItems?.some((item) => item?.orderType === "NOTICE")) ||
-            orderData?.orderType === "NOTICE"
-          ) {
-            history.push(
-              `/${window?.contextPath}/employee/orders/generate-order?filingNumber=${caseDetails?.filingNumber}&orderNumber=${orderData.orderNumber}`,
-              {
-                caseId: caseId,
-                tab: "Orders",
-              }
-            );
-          } else {
-            handleIssueNotice(hearingDate, hearingNumber);
-          }
-        }
-        break;
-      case "SCHEDULE_ADMISSION_HEARING":
-        setShowScheduleHearingModal(true);
-        setCreateAdmissionOrder(true);
-        break;
-
-      default:
-        break;
-    }
-  };
-
-  const onSaveDraft = async () => {
-    if ([CaseWorkflowState.PENDING_NOTICE, CaseWorkflowState.PENDING_RESPONSE].includes(caseDetails?.status)) {
-      const { hearingDate, hearingNumber } = await getHearingData();
-      if (hearingNumber) {
-        const date = new Date(hearingDate);
-        const requestBody = {
-          order: {
-            createdDate: null,
-            tenantId: tenantId,
-            hearingNumber: hearingNumber,
-            filingNumber: filingNumber,
-            cnrNumber: cnrNumber,
-            statuteSection: {
-              tenantId: tenantId,
-            },
-            orderTitle: "INITIATING_RESCHEDULING_OF_HEARING_DATE",
-            orderCategory: "INTERMEDIATE",
-            orderType: "INITIATING_RESCHEDULING_OF_HEARING_DATE",
-            status: "",
-            isActive: true,
-            workflow: {
-              action: OrderWorkflowAction.SAVE_DRAFT,
-              comments: "Creating order",
-              assignes: null,
-              rating: null,
-              documents: [{}],
-            },
-            documents: [],
-            additionalDetails: {
-              formdata: {
-                orderType: {
-                  type: "INITIATING_RESCHEDULING_OF_HEARING_DATE",
-                  isactive: true,
-                  code: "INITIATING_RESCHEDULING_OF_HEARING_DATE",
-                  name: "ORDER_TYPE_INITIATING_RESCHEDULING_OF_HEARING_DATE",
-                },
-                originalHearingDate: `${date.getFullYear()}-${date.getMonth() < 9 ? `0${date.getMonth() + 1}` : date.getMonth() + 1}-${
-                  date.getDate() < 10 ? `0${date.getDate()}` : date.getDate()
-                }`,
-              },
-            },
-          },
-        };
-        ordersService
-          .createOrder(requestBody, { tenantId: Digit.ULBService.getCurrentTenantId() })
-          .then((res) => {
-            history.push(`/${window.contextPath}/employee/orders/generate-order?filingNumber=${filingNumber}&orderNumber=${res.order.orderNumber}`, {
-              caseId: caseDetails?.id,
-              tab: "Orders",
-            });
-          })
-          .catch((err) => {
-            console.error("Error while creating order", err);
-            showToast({ isError: true, message: "ORDER_CREATION_FAILED" });
-          });
-      }
-    } else {
-      setShowModal(true);
-      setSubmitModalInfo({
-        ...scheduleCaseSubmitConfig,
-        caseInfo: [...caseInfo],
-        shortCaseInfo: [
-          {
-            key: "CASE_NUMBER",
-            value: caseDetails?.caseNumber,
-          },
-          {
-            key: "COURT_NAME",
-            value: t(`COMMON_MASTERS_COURT_R00M_${caseCourtId}`),
-          },
-          {
-            key: "CASE_TYPE",
-            value: "NIA S138",
-          },
-        ],
-      });
-      setModalInfo({ type: "schedule", page: 0 });
-    }
-  };
-
   const handleMakeSubmission = useCallback(() => {
     history.push(`/${window?.contextPath}/citizen/submissions/submissions-create?filingNumber=${filingNumber}`);
   }, [filingNumber, history]);
@@ -2546,17 +2418,121 @@ const AdmittedCaseV2 = () => {
     }
   }, [caseDetails, downloadPdf, tenantId, showToast]);
 
+  const pipComplainants = useMemo(() => {
+    return caseDetails?.litigants
+      ?.filter((litigant) => litigant.partyType.includes("complainant"))
+      ?.filter(
+        (litigant) =>
+          !caseDetails?.representatives?.some((representative) =>
+            representative?.representing?.some((rep) => rep?.individualId === litigant?.individualId)
+          )
+      );
+  }, [caseDetails]);
+
+  const pipAccuseds = useMemo(() => {
+    return caseDetails?.litigants
+      ?.filter((litigant) => litigant.partyType.includes("respondent"))
+      ?.filter(
+        (litigant) =>
+          !caseDetails?.representatives?.some((representative) =>
+            representative?.representing?.some((rep) => rep?.individualId === litigant?.individualId)
+          )
+      );
+  }, [caseDetails]);
+
+  const complainantsList = useMemo(() => {
+    const loggedinUserUuid = userInfo?.uuid;
+    // If logged in person is an advocate
+    const isAdvocateLoggedIn = caseDetails?.representatives?.find((rep) => rep?.additionalDetails?.uuid === loggedinUserUuid);
+    const isPipLoggedIn = pipComplainants?.find((p) => p?.additionalDetails?.uuid === loggedinUserUuid);
+    const accusedLoggedIn = pipAccuseds?.find((p) => p?.additionalDetails?.uuid === loggedinUserUuid);
+
+    if (isAdvocateLoggedIn) {
+      return isAdvocateLoggedIn?.representing?.map((r) => {
+        return {
+          code: r?.additionalDetails?.fullName,
+          name: r?.additionalDetails?.fullName,
+          uuid: r?.additionalDetails?.uuid,
+        };
+      });
+    } else if (isPipLoggedIn) {
+      return [
+        {
+          code: isPipLoggedIn?.additionalDetails?.fullName,
+          name: isPipLoggedIn?.additionalDetails?.fullName,
+          uuid: isPipLoggedIn?.additionalDetails?.uuid,
+        },
+      ];
+    } else if (accusedLoggedIn) {
+      return [
+        {
+          code: accusedLoggedIn?.additionalDetails?.fullName,
+          name: accusedLoggedIn?.additionalDetails?.fullName,
+          uuid: accusedLoggedIn?.additionalDetails?.uuid,
+        },
+      ];
+    }
+    return [];
+  }, [caseDetails, pipComplainants, pipAccuseds, userInfo]);
+
   const handleCitizenAction = useCallback(
-    (option) => {
-      if (option.value === "RAISE_APPLICATION") {
-        history.push(`/${window?.contextPath}/citizen/submissions/submissions-create?filingNumber=${filingNumber}`);
-      } else if (option.value === "SUBMIT_DOCUMENTS") {
-        history.push(`/${window?.contextPath}/citizen/submissions/submit-document?filingNumber=${filingNumber}`);
-      } else if (option.value === "GENERATE_BAIL_BOND") {
-        history.push(`/${window?.contextPath}/citizen/submissions/bail-bond?filingNumber=${filingNumber}`);
+    async (option) => {
+      try {
+        if (option.value === "RAISE_APPLICATION") {
+          history.push(`/${window?.contextPath}/citizen/submissions/submissions-create?filingNumber=${filingNumber}`);
+        } else if (option.value === "SUBMIT_DOCUMENTS") {
+          history.push(`/${window?.contextPath}/citizen/submissions/submit-document?filingNumber=${filingNumber}`);
+        } else if (option.value === "GENERATE_BAIL_BOND") {
+          if (complainantsList?.length === 1) {
+            setApiCalled(true);
+            const res = await DRISTIService?.getPendingTaskService(
+              {
+                SearchCriteria: {
+                  tenantId,
+                  moduleName: "Pending Tasks Service",
+                  moduleSearchCriteria: {
+                    isCompleted: false,
+                    ...(isCitizen && { assignedTo: userInfo?.uuid }),
+                    ...(courtId && { courtId }),
+                    filingNumber,
+                    entityType: "bail bond",
+                  },
+                  limit: 1000,
+                  offset: 0,
+                },
+              },
+              { tenantId }
+            );
+            const pendingTaskResponse = res?.data || [];
+            const pendingTaskDetails = convertTaskResponseToPayload(pendingTaskResponse);
+
+            if (pendingTaskResponse?.length > 0 && pendingTaskDetails?.additionalDetails?.bailbondId) {
+              history.push(
+                `/${window?.contextPath}/citizen/submissions/bail-bond/view?filingNumber=${filingNumber}&bailBondId=${pendingTaskDetails?.additionalDetails?.bailbondId}`
+              );
+            } else if (pendingTaskResponse?.length > 0) {
+              history.push(`/${window?.contextPath}/citizen/submissions/bail-bond?filingNumber=${filingNumber}`, {
+                state: {
+                  params: {
+                    actualReferenceId: pendingTaskDetails?.referenceId,
+                  },
+                },
+              });
+            } else {
+              history.push(`/${window?.contextPath}/citizen/submissions/bail-bond?filingNumber=${filingNumber}`);
+            }
+          } else {
+            history.push(`/${window?.contextPath}/citizen/submissions/bail-bond?filingNumber=${filingNumber}`);
+          }
+        }
+      } catch (error) {
+        console.error("Error handling citizen action:", error);
+        showToast({ isError: true, message: "BAIL_BOND_SEARCH_FAILED" });
+      } finally {
+        setApiCalled(false);
       }
     },
-    [filingNumber, history]
+    [history, filingNumber]
   );
 
   const handleCourtAction = useCallback(() => {
