+++ frontend/micro-ui/web/micro-ui-internals/packages/modules/dristi/src/pages/employee/AdmittedCases/DocumentsV2.js
@@ -242,7 +242,6 @@ const DocumentsV2 = ({
                 criteria: {
                   ...(tabConfig.apiDetails?.requestBody?.criteria || {}),
                   filingNumber: filingNumber,
-                  status: ["DRAFT_IN_PROGRESS", "COMPLETED"],
                 },
               },
             },
