+++ frontend/micro-ui/web/micro-ui-internals/packages/modules/dristi/src/services/index.js
@@ -585,4 +585,12 @@ export const DRISTIService = {
       params,
     });
   },
+  searchBailBonds: (data, params) =>
+    Request({
+      url: Urls.dristi.searchBailBonds,
+      useCache: false,
+      userService: true,
+      data,
+      params,
+    }),
 };
