+++ frontend/micro-ui/web/micro-ui-internals/packages/modules/home/src/components/TaskComponent.js
@@ -663,6 +663,7 @@ const TasksComponent = ({
           dateOfApplication,
           uniqueId,
           applicationType,
+          actualReferenceId: referenceId,
         },
         isCustomFunction,
         referenceId,
