+++ frontend/micro-ui/web/micro-ui-internals/packages/modules/submissions/src/hooks/services/index.js
@@ -130,4 +130,12 @@ export const submissionService = {
       data,
       params,
     }),
+  getPendingTask: (data, params) =>
+    Request({
+      url: Urls.getPendingTaskFields,
+      useCache: false,
+      userService: false,
+      data,
+      params,
+    }),
 };
